Tmp commit

+++ DataToBeProcessed/FileData/datfiledata.cpp
@@ -1,13 +1,13 @@
 #include "datfiledata.h"
 
-DatFileData::DatFileData(DataSourceType sourceType, QString& filePath, std::vector<int> data):
+DatFileData::DatFileData(DataSourceType sourceType, QString& filePath, std::vector<double> data):
     FileData(sourceType, filePath),
     data_(data)
 {
 
 }
 
-std::vector<int> DatFileData::getData()const
+std::vector<double> DatFileData::getData()const
 {
     return data_;
 }
